+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 - refactor(core): make the scheduler compensate for the time it took to run the last scheduled event, by waiting less time in the next scheduled event, also sync its frequency with the frequency of the display updates
 - feat(core): handle system signal `USR1` as the signal to initiate `ipdb` only if `DEBUG_MODE_PDB_SIGNAL` is set
 - refactor(system): improve the user experience of the reset button, add a pattern for restarting the app without killing it
+- refactor(core): add unified API to access “current service” via thread-local, parent thread, or call stack
 
 ## Version 1.2.2